Remove user and rate limit code to a separate file
